TST Split tensorflow 2d tests into backend and scattering. added fft type tests
